(pure-move) migrate allocate_device_buffer and allocate_host_buffer
